Move Flyout.tsx into components dir
